Reorganized headers to be more independent.  Formatted code.
