+++ include/includize/toml.hpp
@@ -2,17 +2,17 @@
 #define INCLUDIZE_TOML_HPP
 
 #include "includize.hpp"
+#include "null_stream_preparer.hpp"
 #include <regex>
 
 namespace includize
 {
-    
-template< typename CHAR_TYPE >
+template < typename CHAR_TYPE >
 struct toml_spec
 {
 };
 
-template<>
+template <>
 struct toml_spec< char >
 {
     static constexpr char header_start() { return '#'; }
@@ -20,60 +20,26 @@ struct toml_spec< char >
     {
         return R"..(\s*\[\[include\s*"(([^"]|\")+)"\s*]])..";
     }
-    
+
     static constexpr std::size_t file_name_index() { return 1; };
     static constexpr bool discard_characters_after_include() { return true; }
-    
+
+    static std::string convert_filename(const std::string &str) { return str; }
+
     static std::string unescape_filename(const std::string &str)
     {
         return std::regex_replace(str, std::regex("\\\\\""), "\"");
     }
 };
 
-#if !defined(INCLUDIZE_NO_CODECVT)
-
-template<>
-struct toml_spec< wchar_t >
-{
-    static constexpr wchar_t header_start() { return L'#'; }
-    static constexpr const wchar_t *regex()
-    {
-        return LR"..(\s*\[\[include\s*"(([^"]|\")+)"\s*]])..";
-    }
-    
-    static constexpr std::size_t file_name_index() { return 1; };
-    static constexpr bool discard_characters_after_include() { return true; }
-    
-    static std::string unescape_filename(const std::string &str)
-    {
-        return std::regex_replace(str, std::regex("\\\\\""), "\"");
-    }
-};
-
-#endif
-
-template< typename CHAR_T, typename TRAITS = std::char_traits< CHAR_T > >
-using basic_toml_streambuf = basic_streambuf< toml_spec< CHAR_T >, CHAR_T, TRAITS >;
-
-using toml_streambuf = basic_toml_streambuf< char >;
-
-#if !defined(INCLUDIZE_NO_CODECVT)
-
-using wtoml_streambuf = basic_toml_streambuf< wchar_t >;
-
-#endif
-
-template< typename CHAR_T, typename TRAITS = std::char_traits< CHAR_T > >
-using basic_toml_preprocessor = basic_preprocessor< toml_spec< CHAR_T >, CHAR_T, TRAITS >;
+template < typename CHAR_T,
+           typename TRAITS = std::char_traits< CHAR_T >,
+           typename STREAM_PREPARER = null_stream_preparer< CHAR_T, TRAITS > >
+using basic_toml_preprocessor =
+    basic_preprocessor< toml_spec< CHAR_T >, CHAR_T, TRAITS, STREAM_PREPARER >;
 
 using toml_preprocessor = basic_toml_preprocessor< char >;
 
-#if !defined(INCLUDIZE_NO_CODECVT)
-
-using wtoml_preprocessor = basic_toml_preprocessor< wchar_t >;
-
-#endif
-    
-} // namespace includize
+}  // namespace includize
 
 #endif
