+++ tools/format-code
@@ -1,0 +1,68 @@
+#!/bin/bash
+
+# Copyright (c) 2015, Daniel C. Dillon
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice,
+#    this list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright
+#    notice, this list of conditions and the following disclaimer in the
+#    documentation and/or other materials provided with the distribution.
+#
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+# TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+# PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+# PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+# LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+# NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+pushd $DIR/.. > /dev/null 2>&1
+
+FILES=`find -name *.hpp`
+
+for FILE in $FILES ; do
+	BASENAME=`basename $FILE`
+
+	if [ "$BASENAME" != "catch.hpp" ]; then
+		clang-format -style=file -i $FILE
+	fi
+done
+
+FILES=`find -name *.h`
+
+for FILE in $FILES ; do
+	BASENAME=`basename $FILE`
+
+	if [ "$BASENAME" != "cpptoml.h" ]; then
+		clang-format -style=file -i $FILE
+	fi
+done
+
+FILES=`find -name *.cpp`
+
+for FILE in $FILES ; do
+	clang-format -style=file -i $FILE
+done
+
+FILES=`find -name *.inl`
+
+for FILE in $FILES ; do
+	clang-format -style=file -i $FILE
+done
+
+popd > /dev/null 2>&1
